pyproject.toml
[tool.poetry]
name = "gcp_data_preparation"
version = "0.1.0"
description = ""
authors = ["jihoahn9303 <rediate03@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
hydra-core = "^1.3.2"
pydantic = "^2.8.2"
pandas = "~=2.0"
nltk = "~=3.8"
symspellpy = "~=6.7"
fsspec = {version = "~=2023.5", extras = ["gcs"]}
gcsfs = "~=2023.5"
google-cloud-secret-manager = "~2.20.2"
fastparquet = "~=2023.4"
dvc = {version = "~=2.56", extras = ["gdrive", "gs"]}
dask = {version = "~=2024.8", extras = ["complete"]}
distributed = "~=2024.8"
dask-cloudprovider = {extras = ["gcp"], version = "^2022.10.0"}
dask-ml = "~=2023.3"
pygit2 = "1.14.1"
jupyterlab = "^4.2.4"
bokeh = "~=3.5"
toolz = "0.12.0"

# Tokenization
tokenizers = "~=0.13"
transformers = "~=4.30.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

types-PyYAML = "~=6.0.12.20240808"
types-cachetools = "~=5.4.0.20240717"
types-decorator = "~=5.1.8.20240310"
types-protobuf = "~=5.27.0.20240626"
types-pycurl = "~=7.45.3.20240421"
types-requests = "~=2.32.0.20240712" 
types-six = "~=1.16.21.20240513"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party = ["jeffrey"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.9"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true

disallow_subclassing_any = true 
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0
